renderutils refactor & hud gui pre-update pre-beta v4

+++ Firework-Client/src/main/java/com/firework/client/Implementations/Utill/Render/Rectangle.java
@@ -2,12 +2,12 @@ package com.firework.client.Implementations.Utill.Render;
 
 public class Rectangle {
 
-    public int x;
-    public int y;
+    public double x;
+    public double y;
     public double width;
     public double height;
 
-    public Rectangle(int x, int y, double width, double height){
+    public Rectangle(double x, double y, double width, double height){
         this.x = x;
         this.y = y;
         this.width = width;
